requirements.txt

click==8.1.3
pillow==9.5.0
requests==2.28.2
httpx==0.24.0
pyyaml==6.0
# Development dependencies
flake8==3.9.2
black==21.6b0
pytest==6.2.5
jsonschema